*: add 'psn ds' command
