+++ .travis.yml
@@ -8,5 +8,4 @@ go:
 script:
 - ./test
 - go get -v github.com/gyuho/psn/cmd/psn
-- psn ps
-- psn ss
+- psn ds
