+++ cmd/psn/main.go
@@ -1,21 +1,3 @@
-// psn provides utilities to investigate OS processes and sockets.
-//
-//	Usage:
-//	  psn [command]
-//
-//	Available Commands:
-//	  process         Investigates processes status
-//	  process-kill    Kills processes
-//	  process-monitor Monitors processes
-//	  socket          Investigates sockets
-//	  socket-kill     Kills sockets
-//	  socket-monitor  Monitors sockets
-//
-//	Flags:
-//	  -h, --help[=false]: help for psn
-//
-//	Use "psn [command] --help" for more information about a command.
-//
 package main
 
 import (
@@ -28,18 +10,13 @@ import (
 var (
 	Command = &cobra.Command{
 		Use:        "psn",
-		Short:      "psn provides utilities to investigate OS processes and sockets.",
+		Short:      "psn inspects Linux processes, sockets (ps, ss, netstat).",
 		SuggestFor: []string{"pssn", "psns", "snp"},
 	}
 )
 
 func init() {
-	Command.AddCommand(processCommand)
-	Command.AddCommand(processKillCommand)
-	Command.AddCommand(processMonitorCommand)
-	Command.AddCommand(socketCommand)
-	Command.AddCommand(socketKillCommand)
-	Command.AddCommand(socketMonitorCommand)
+	Command.AddCommand(dsCommand)
 }
 
 func init() {
